Parsing without expressions most likely done
Added a bunch of unit tests, and it seems that parsing and matching of
simple patterns is fully operational. May remake the API to make it more
dev-friendly.
Now, expression parsing is missing altogether, and regex has not been
tested yet, even though I'm not worried that much about it.