Improved registration regarding types
SkriptRegistration is structured in such a way that adding a new syntax
does not make it usable right away, you have to register before doing
that. This is problematic with types, since they are required for
pattern parsing, contrary to all other registerable syntaxes.
This change removes the need to manually register types.